Modified README.md and other small improvements

+++ unity/GpuParticleSimulations/Assets/SimulationEngine/Settings/BoidsSettings.cs
@@ -14,6 +14,6 @@ namespace SimulationEngine.Settings
         public float centeringFactor = 1;
         public float avoidFactor = 1;
         public float turnFactor = 0.1f;
-        public float boidsEnemies = 10;
+        public int boidsEnemies = 10;
     }
 }
